lazy transfer

+++ main.c
@@ -263,10 +263,10 @@ char ValidateChecksum()
 {
 	unsigned long newChkSum = CRC32_Tab(msgToRelay, byteCntToRead, -1);
 
-	unsigned long control = checksum[3];
-	control = control | checksum[0] << 8 * 3;
-	control = control | checksum[1] << 8 * 2;
-	control = control | checksum[2] << 8 * 1;
+	unsigned long control = (unsigned long)checksum[3];
+	control = control | (unsigned long)checksum[0] << 8 * 3;
+	control = control | (unsigned long)checksum[1] << 8 * 2;
+	control = control | (unsigned long)checksum[2] << 8 * 1;
 
 	TerminalWriteText("newChecksum\n");
 	TerminalWrite(newChkSum >> 8 * 3);
